# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Zensor Server API
  description: Zensor Server API for IoT device management and task scheduling
  version: 1.0.0
  contact:
    name: Zensor Team
    email: support@zensor-iot.net
  license:
    name: Proprietary
    url: https://zensor-iot.net
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.zensor-iot.net
    description: Production server
paths:
...
  /v1/tenants/{tenant_id}/devices/{device_id}/scheduled-tasks:
    get:
      summary: List scheduled tasks
      description: Retrieve all scheduled tasks for a specific tenant and device
      tags:
        - Scheduled Tasks
      parameters:
        - name: tenant_id
          in: path
          required: true
          description: Tenant ID
          schema:
            type: string
            format: uuid
        - name: device_id
          in: path
          required: true
          description: Device ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: List of scheduled tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedScheduledTaskResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create scheduled task
      description: Create a new scheduled task for a specific tenant and device
      tags:
        - Scheduled Tasks
      parameters:
        - name: tenant_id
          in: path
          required: true
          description: Tenant ID
          schema:
            type: string
            format: uuid
        - name: device_id
          in: path
          required: true
          description: Device ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledTaskCreateRequest'
      responses:
        '201':
          description: Scheduled task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledTaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/tenants/{tenant_id}/devices/{device_id}/scheduled-tasks/{id}:
    get:
      summary: Get scheduled task
      description: Retrieve a specific scheduled task
      tags:
        - Scheduled Tasks
      parameters:
        - name: tenant_id
          in: path
          required: true
          description: Tenant ID
          schema:
            type: string
            format: uuid
        - name: device_id
          in: path
          required: true
          description: Device ID
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Scheduled task ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Scheduled task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledTaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      summary: Update scheduled task
      description: Update an existing scheduled task
      tags:
        - Scheduled Tasks
      parameters:
        - name: tenant_id
          in: path
          required: true
          description: Tenant ID
          schema:
            type: string
            format: uuid
        - name: device_id
          in: path
          required: true
          description: Device ID
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Scheduled task ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduledTaskUpdateRequest'
      responses:
        '200':
          description: Scheduled task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduledTaskResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      summary: Soft delete scheduled task
      description: Soft delete a scheduled task (marks as deleted but doesn't remove from database)
      tags:
        - Scheduled Tasks
      parameters:
        - name: tenant_id
          in: path
          required: true
          description: Tenant ID
          schema:
            type: string
            format: uuid
        - name: device_id
          in: path
          required: true
          description: Device ID
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Scheduled task ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Scheduled task soft deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Scheduled task is already soft deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/tenants/{tenant_id}/devices/{device_id}/scheduled-tasks/{id}/tasks:
    get:
      summary: Get tasks by scheduled task
      description: Retrieve all tasks created by a specific scheduled task
      tags:
        - Scheduled Tasks
      parameters:
        - name: tenant_id
          in: path
          required: true
          description: Tenant ID
          schema:
            type: string
            format: uuid
        - name: device_id
          in: path
          required: true
          description: Device ID
          schema:
            type: string
            format: uuid
        - name: id
          in: path
          required: true
          description: Scheduled task ID
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Page number for pagination
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        '200':
          description: List of tasks created by the scheduled task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTaskResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
      example:
        message: "An error occurred"

    # Tenant schemas
...
    PaginationInfo:
      type: object
      properties:
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Number of items per page
          example: 10
        total:
          type: integer
          description: Total number of items
          example: 100
        total_pages:
          type: integer
          description: Total number of pages
          example: 10
...
    CommandSendPayloadRequest:
      type: object
      required:
        - index
        - value
      properties:
        wait_for:
          type: string
          description: Duration to wait before executing this command
          example: "30s"
        priority:
          type: string
          description: Command priority
          enum: [LOW, NORMAL, HIGH]
          example: "NORMAL"
        index:
          type: integer
          minimum: 0
          maximum: 255
          description: Sensor/actuator index
          example: 1
        value:
          type: integer
          minimum: 0
          maximum: 255
          description: Command value
          example: 100
...
    TaskCommandResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Command ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        index:
          type: integer
          minimum: 0
          maximum: 255
          description: Sensor/actuator index
          example: 1
        value:
          type: integer
          minimum: 0
          maximum: 255
          description: Command value
          example: 100
        port:
          type: integer
          minimum: 0
          maximum: 255
          description: Command port
          example: 1
        priority:
          type: string
          description: Command priority
          example: "NORMAL"
        dispatch_after:
          type: string
          format: date-time
          description: When the command should be dispatched
          example: "2024-01-01T00:00:00Z"
        ready:
          type: boolean
          description: Whether the command is ready to be sent
          example: true
        sent:
          type: boolean
          description: Whether the command has been sent
          example: false
        sent_at:
          type: string
          format: date-time
          nullable: true
          description: When the command was sent
          example: "2024-01-01T00:00:00Z"
    TaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Task ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        commands:
          type: array
          items:
            $ref: '#/components/schemas/TaskCommandResponse'
          description: Commands in the task
        created_at:
          type: string
          format: date-time
          description: Task creation timestamp
          example: "2024-01-01T00:00:00Z"
...
    ScheduledTaskCreateRequest:
      type: object
      required:
        - commands
        - schedule
      properties:
        commands:
          type: array
          items:
            $ref: '#/components/schemas/CommandSendPayloadRequest'
          description: Commands to execute as part of the scheduled task
        schedule:
          type: string
          description: Cron expression for scheduling
          example: "0 */5 * * * *"
        is_active:
          type: boolean
          description: Whether the scheduled task is active
          default: true
          example: true
    ScheduledTaskUpdateRequest:
      type: object
      properties:
        commands:
          type: array
          items:
            $ref: '#/components/schemas/CommandSendPayloadRequest'
          description: Commands to execute as part of the scheduled task
        schedule:
          type: string
          description: Cron expression for scheduling
          example: "0 */5 * * * *"
        is_active:
          type: boolean
          description: Whether the scheduled task is active
          example: true
    ScheduledTaskResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Scheduled task ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        device_id:
          type: string
          format: uuid
          description: Target device ID
          example: "123e4567-e89b-12d3-a456-426614174000"
        commands:
          type: array
          items:
            $ref: '#/components/schemas/CommandSendPayloadRequest'
          description: Commands in the scheduled task
        schedule:
          type: string
          description: Cron expression for scheduling
          example: "0 */5 * * * *"
        is_active:
          type: boolean
          description: Whether the scheduled task is active
          example: true
    PaginatedScheduledTaskResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ScheduledTaskResponse'
          description: Array of scheduled task objects
        pagination:
          $ref: '#/components/schemas/PaginationInfo'

    # Evaluation Rule schemas
...
    PaginatedTaskResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/TaskResponse'
          description: Array of task objects
        pagination:
          $ref: '#/components/schemas/PaginationInfo'
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...